added notification schudeler generator

+++ src/main/java/com/UserApi/Repository/UserRepo.java
@@ -3,7 +3,8 @@ package com.UserApi.Repository;
 import com.UserApi.Entities.UserDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<UserDetails, Integer>{
+public interface UserRepo extends JpaRepository<UserDetails, Long>{
+
